Koin DI added to handle viewModels, screens now retains selected user info on recompose
