+++ app/src/main/java/com/davidspartan/androidflipcardgame/MainActivity.kt
@@ -1,16 +1,20 @@
 package com.davidspartan.androidflipcardgame
 
-import android.content.pm.ActivityInfo
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
+import androidx.compose.foundation.layout.Box
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
 import androidx.core.view.WindowCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.WindowInsetsControllerCompat
+import androidx.lifecycle.ViewModel
 import androidx.navigation.compose.rememberNavController
 import com.davidspartan.androidflipcardgame.view.navigation.NavHost
 import com.davidspartan.androidflipcardgame.viewmodel.UserRepositoryViewModel
+import org.koin.androidx.compose.koinViewModel
 
 
 
@@ -19,25 +23,20 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
 
-        // Lock orientation to portrait
-        //requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT
-
         val insetsController = WindowCompat.getInsetsController(window, window.decorView)
 
         insetsController.apply {
             hide(WindowInsetsCompat.Type.navigationBars())
             systemBarsBehavior = WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
         }
+
         setContent {
             val navController = rememberNavController()
-            val viewModel = UserRepositoryViewModel()
 
-            NavHost(navController,viewModel)
+            NavHost(navController)
         }
     }
 }
-
-
