--- core/database/src/main/java/com/davidspartan/database/realm/MyApp.kt
@@ -1,24 +1,0 @@
-package com.davidspartan.database.realm
-
-import android.app.Application
-import io.realm.kotlin.Realm
-import io.realm.kotlin.RealmConfiguration
-
-class MyApp: Application() {
-
-    companion object {
-        lateinit var realm: Realm
-    }
-
-    override fun onCreate() {
-        super.onCreate()
-        realm = Realm.open(
-            configuration = RealmConfiguration.create(
-                schema = setOf(
-                    User::class,
-                    Theme::class,
-                )
-            )
-        )
-    }
-}
