+++ core/database/src/main/java/com/davidspartan/database/realm/MyRealm.kt
@@ -1,0 +1,12 @@
+package com.davidspartan.database.realm
+
+import android.app.Application
+import io.realm.kotlin.Realm
+
+class MyRealm: Application() {
+
+    companion object {
+        lateinit var realm: Realm
+    }
+
+}
