+++ core/model/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
 <application
-    android:name="com.davidspartan.database.realm.MyApp">
+    android:name="com.davidspartan.database.realm.MyRealm">
 
 </application>
 </manifest>
